flag maintenance, cemetery map filtering

+++ config/cemetery.php
@@ -14,15 +14,15 @@ return [
     */
 
     'center' => [
-        'latitude' => env('CEMETERY_CENTER_LAT', 14.5995),
-        'longitude' => env('CEMETERY_CENTER_LNG', 120.9842),
+        'latitude' => (float) env('CEMETERY_CENTER_LAT', 14.5995),
+        'longitude' => (float) env('CEMETERY_CENTER_LNG', 120.9842),
     ],
 
     'entrance' => [
-        'latitude' => env('CEMETERY_ENTRANCE_LAT', env('CEMETERY_CENTER_LAT', 14.5995)),
-        'longitude' => env('CEMETERY_ENTRANCE_LNG', env('CEMETERY_CENTER_LNG', 120.9842)),
+        'latitude' => (float) env('CEMETERY_ENTRANCE_LAT', env('CEMETERY_CENTER_LAT', 14.5995)),
+        'longitude' => (float) env('CEMETERY_ENTRANCE_LNG', env('CEMETERY_CENTER_LNG', 120.9842)),
     ],
 
-    'zoom' => env('CEMETERY_ZOOM', 16),
+    'zoom' => (int) env('CEMETERY_ZOOM', 16),
 
 ];
